+++ tests/test_parsing.py
@@ -1,48 +1,31 @@
 """Module for tests related to parsing."""
 import json
+import re
 
 import pytest
-
 from beetsplug.bandcamp._metaguru import NEW_BEETS, Metaguru, urlify
+from pytest_lazyfixture import lazy_fixture
 
 pytestmark = pytest.mark.parsing
 
 
-def check(actual, expected) -> None:
-    if NEW_BEETS:
-        assert actual == expected
-    else:
-        assert vars(actual) == vars(expected)
-
-
 @pytest.mark.parametrize(
-    ("meta", "expected"),
+    ("descr", "disctitle", "creds", "expected"),
     [
-        ({}, ""),
-        ({"description": "hello"}, "hello"),
-        ({"description": "Includes high-quality download"}, ""),
-        (
-            {
-                "description": "",
-                "albumRelease": [
-                    {"musicReleaseFormat": "VinylFormat", "description": "sick vinyl"}
-                ],
-            },
-            "sick vinyl",
-        ),
-        (
-            {
-                "description": "sickest vinyl",
-                "albumRelease": [
-                    {"musicReleaseFormat": "VinylFormat", "description": "sick vinyl"}
-                ],
-            },
-            "sickest vinyl",
-        ),
+        ("", "", "", ""),
+        ("hello", "", "", "\n - hello"),
+        ("", "Includes high-quality download", "Thanks", "\n - Credits: Thanks"),
+        ("", "sick vinyl", "", "\n - sick vinyl"),
+        ("sickest vinyl", "sick vinyl", "", "\n - sickest vinyl\n - sick vinyl"),
     ],
 )
-def test_description(meta, expected):
-    meta.update(datePublished="doesntmatter")
+def test_description(descr, disctitle, creds, expected):
+    meta = dict(
+        description=descr,
+        albumRelease=[{"musicReleaseFormat": "VinylFormat", "description": disctitle}],
+        creditText=creds,
+        datePublished="doesntmatter",
+    )
     guru = Metaguru(json.dumps(meta), media="Vinyl")
     guru._media = meta.get("albumRelease", [{}])[0]
     assert guru.description == expected, vars(guru)
@@ -104,6 +87,7 @@ def test_parse_release_date(string, expected):
         ("24 Hours", (None, None, "24 Hours")),
         ("Some tune (Someone's Remix)", (None, None, "Some tune (Someone's Remix)")),
         ("19.85 - Colapso Inevitable", (None, "19.85", "Colapso Inevitable")),
+        ("19.85 - Colapso Inevitable (FREE)", (None, "19.85", "Colapso Inevitable")),
     ],
 )
 def test_parse_track_name(name, expected):
@@ -116,17 +100,16 @@ def test_parse_track_name(name, expected):
     [
         ("Artist - Track [Digital Bonus]", True, "Artist - Track"),
         ("DIGI 11. Track", True, "Track"),
-        ("Digital Life", False, None),
+        ("Digital Life", False, "Digital Life"),
         ("Messier 33 (Bandcamp Digital Exclusive)", True, "Messier 33"),
         ("33 (bandcamp exclusive)", True, "33"),
         ("Tune (Someone's Remix) [Digital Bonus]", True, "Tune (Someone's Remix)"),
     ],
 )
 def test_check_digital_only(name, expected_digital_only, expected_name):
-    expected = dict(digital_only=expected_digital_only)
-    if expected_name:
-        expected.update(name=expected_name)
-    assert Metaguru.check_digital_only(name) == expected
+    actual_name, actual_digi_only = Metaguru.clean_digital_only_track(name)
+    assert actual_digi_only == expected_digital_only
+    assert actual_name == expected_name
 
 
 @pytest.mark.parametrize(
@@ -181,6 +164,7 @@ def test_parse_country(name, expected):
         ("Catalogue: CTU-300", "UTC-003", "CTU-300"),
         ("Cat No: TE0029", "UTC-003", "TE0029"),
         ("Cat Nr.: TE0029", "UTC-003", "TE0029"),
+        ("Catalogue:CTU-300", "UTC-003", "CTU-300"),
     ],
 )
 def test_parse_catalognum(description, album, expected):
@@ -215,21 +199,60 @@ def test_parse_catalognum(description, album, expected):
         ("Some feat. Some ONE - Album", ["Some feat. Some ONE"], "Album"),
         ("Healing Noise (EP) (Free Download)", [], "Healing Noise"),
         ("[MCVA003] - VARIOUS ARTISTS", ["MCVA003"], "MCVA003"),
+        ("Drepa Mann [Vinyl]", [], "Drepa Mann"),
     ],
 )
 def test_clean_up_album_name(album, extras, expected):
     assert Metaguru.clean_up_album_name(album, *extras) == expected
 
 
-def test_parse_single_track_release(single_track_release):
-    html, expected = single_track_release
+@pytest.fixture(name="release")
+def _release(request):
+    """Read the json data and make it span a single line - same like it's found in htmls.
+    Include the release date too - we still need to get it directly.
+    Fixture names map to the testfiles (minus the extension).
+    """
+    info = request.param
+    prepend = info.html_release_date + "\n"
+    fixturename = next(iter(request._parent_request._fixture_defs.keys()))
+    filename = "tests/json/{}.json".format(fixturename)
+    return prepend + re.sub(r"\n *", "", open(filename).read()), info
+
+
+def check(actual, expected) -> None:
+    if NEW_BEETS:
+        assert actual == expected
+    else:
+        assert vars(actual) == vars(expected)
+
+
+@pytest.mark.parametrize(
+    "release", [lazy_fixture("single_track_release")], indirect=["release"]
+)
+def test_parse_single_track_release(release):
+    html, expected = release
     guru = Metaguru(html)
 
     check(guru.singleton, expected.singleton)
 
 
-def test_parse_album_or_comp(multitracks):
-    html, expected_release = multitracks
+@pytest.mark.parametrize(
+    "release",
+    map(
+        lazy_fixture,
+        [
+            "album",
+            "album_with_track_alt",
+            "compilation",
+            "ep",
+            "artist_mess",
+            "description_meta",
+        ],
+    ),
+    indirect=["release"],
+)
+def test_parse_various_types(release):
+    html, expected_release = release
     guru = Metaguru(html, expected_release.media)
     include_all = False